refactor(ui): fast-refresh compliance batch 3 — 8 single-component route files
Part of #7850. 8 of 246 react-refresh/only-export-components warnings.

The dominant pattern behind the 33-file route-tree share of this sweep:
export const Route = createFileRoute(...)({ component: XPage, ... }) --
XPage is referenced from the exported Route config but never itself
exported, so fast-refresh can't tell the file is component-only. Fixed
identically across all 8: moved the page component (and everything only
it uses) to a new `-<route>-page.tsx` sibling (the `-` prefix matches
the existing routeFileIgnorePrefix convention, so TanStack's file-based
router ignores it); each route file now exports only Route.

Two needed the closure-over-Route pattern swapped for the equivalent
standalone hook (same as batch 2's RootComponent), since moving the
component out of the route file would make importing Route back in
circular:
- events.index.tsx: Route.useSearch()/Route.fullPath -> useSearch({from})/useNavigate({from})
- docs.$.tsx: Route.useLoaderData()/Route.useParams() -> useLoaderData({from})/useParams({from})
  (also moved clientLoader, Page's only consumer, into the new sibling
  file alongside it -- same split-chunk boundary either way, since
  what matters for TanStack's code-splitting is the `component:` field,
  not which physical file defines the referenced function, per the
  file's own existing comment about why loader can't reference it)

index.tsx additionally had a second, unrelated export --
ErrorBoundaryFallback, a real component, but dead code (grep confirms
zero importers anywhere, not wired to any errorComponent). Left it
exactly as dead code and moved it into the new -index-page.tsx sibling
along with the ~13 other local components this route already defines --
not this PR's job to prune it, just to stop it from coexisting with
Route in a way that trips the rule.

Zero behavior change; no external importers of any of the 8 moved
components existed outside their own route file, so this is fully
mechanical.

246 -> 238 warnings; 0 eslint errors. Live-verified all 8 routes
(homepage, /events, /docs/graphql, /graphql/explorer, /domains,
/accounts, /portfolio, /settings) -- render correctly, no console
errors.
