+++ apps/ui/src/routes/domains.tsx
@@ -1,10 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router";
-import { AppShell } from "@/components/metagraphed/app-shell";
-import { ApiSourceFooter } from "@/components/metagraphed/api-source-footer";
-import { DomainsRollup } from "@/components/metagraphed/domains-rollup";
-import { ActionBar, ShareButton } from "@jsonbored/ui-kit";
-import { AsyncPanel, PageMasthead, PanelSkeleton } from "@/components/metagraphed/primitives";
-import { metagraphedQueryKey } from "@/lib/metagraphed/queries";
+import { DomainsPage } from "./-domains-page";
 
 export const Route = createFileRoute("/domains")({
   head: () => ({
@@ -25,29 +20,3 @@ export const Route = createFileRoute("/domains")({
   }),
   component: DomainsPage,
 });
-
-function DomainsPage() {
-  return (
-    <AppShell>
-      <PageMasthead
-        eyebrow="Explorer"
-        live
-        title="Domains"
-        description="The 14-tag capability taxonomy — every domain with its member subnets, total stake, emission share, and within-domain emission concentration. Expand a domain to see its full concentration breakdown and jump to any member subnet."
-        actions={
-          <ActionBar>
-            <ShareButton bare />
-          </ActionBar>
-        }
-      />
-      <AsyncPanel
-        context="domains"
-        fallback={<PanelSkeleton height="md" />}
-        retryQueryKeys={[metagraphedQueryKey("domains"), metagraphedQueryKey("subnets")]}
-      >
-        <DomainsRollup />
-      </AsyncPanel>
-      <ApiSourceFooter paths={["/api/v1/domains", "/api/v1/subnets"]} />
-    </AppShell>
-  );
-}
